requirements.txt
flask
deepface
tensorflow==2.20.0
tf-keras
pillow
numpy
